requirements.txt
transformers
huggingface
pandas
pillow 
datasets 
torch
pdf2image
protobuf
sentencepiece
torchvision